# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Backend Boilerplate",
        "version": "1.0"
    },
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/roles": {
            "get": {
                "tags": [
                    "roles"
                ],
                "summary": "Get all roles",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/Role Response"
                        }
                    },
                    "422": {
                        "description": "error details"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/roles/{id}": {
            "put": {
                "tags": [
                    "roles"
                ],
                "summary": "Update a role",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/Role Response"
                        }
                    },
                    "422": {
                        "description": "error details"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "roles"
                ],
                "summary": "Create new role",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/Role Response"
                        }
                    },
                    "422": {
                        "description": "error details"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "roles"
                ],
                "summary": "Delete a role",
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/Role Response"
                        }
                    },
                    "422": {
                        "description": "error details"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "title": "User",
                            "properties": {
                                "first_name": {
                                    "type": "string"
                                },
                                "last_name": {
                                    "type": "string"
                                },
                                "username": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "role": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "422": {
                        "description": "error details"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "User's Profile",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "422": {
                        "description": "error details"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "Role Response": {
            "title": "Role",
            "properties": {
                "id": {
                    "description": "UUID",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "guard_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "Role": {
            "title": "Role Request",
            "required": [
                "name",
                "guard_name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "guard_name": {
                    "type": "string"
                }
            }
        },
        "User": {
            "title": "User",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "roles": {
                    "items": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "guard_name": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "updated_at": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "type": "object"
        }
    },
    "securityDefinitions": {
        "implicit": {
            "type": "oauth2",
            "in": "header",
            "flow": "implicit"
        }
    }
}